refactor: App.tsx 에 있는 header 라우팅 함수 훅으로 분리 header 폴더 안에 넣어서 여기서만 쓸 예정
